backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
python-pptx==0.6.23
pdfplumber==0.11.0
deep-translator==1.11.4
werkzeug==3.0.3
gunicorn==21.2.0
